Comment on OpenGL import version
When I made this app, I assumed the user would be using OpenGL 3.2 or greater.  However, should you be using an older version of OpenGL (2.1 or less), you will need to add or exchange 'import OpenGL.GL3' for 'import OpenGL.GL'.  You should not need to import OpenGL.GL if you are staying in the a modern OpenGL paradigm (see the Red Book covering 3.2 or higher or The Super Bible covering 3.2 or higher to brush up on using modern OpenGL).